Update to 1.21.3
Notable changes included in this commit:
- Pre-world initialization is now performed during bootstrap phase.
- ArmorRegistry has been removed. Instead, the new EquipmentRegistry takes advantage of the new minecraft:equippable data component. There are now dsl builders for creating animations for equipment textures.
- Removed FancyPants armor shader and replaced it with a simplified version that only handles emissivity. Also removed FancyPants armor file merger.
- Added new functionality to Consumable behavior
- Added Cooldown item behavior

+++ nova/src/main/kotlin/xyz/xenondevs/nova/patch/impl/worldgen/chunksection/LevelChunkSectionWrapper.java
@@ -61,8 +61,8 @@ public class LevelChunkSectionWrapper extends LevelChunkSection {
             SET_TICKING_BLOCK_COUNT = lookup.findSetter(LevelChunkSection.class, "tickingBlockCount", short.class);
             GET_TICKING_FLUID_COUNT = lookup.findGetter(LevelChunkSection.class, "tickingFluidCount", short.class);
             SET_TICKING_FLUID_COUNT = lookup.findSetter(LevelChunkSection.class, "tickingFluidCount", short.class);
-            GET_SPECIAL_COLLIDING_BLOCKS = lookup.findGetter(LevelChunkSection.class, "specialCollidingBlocks", int.class);
-            SET_SPECIAL_COLLIDING_BLOCKS = lookup.findSetter(LevelChunkSection.class, "specialCollidingBlocks", int.class);
+            GET_SPECIAL_COLLIDING_BLOCKS = lookup.findGetter(LevelChunkSection.class, "specialCollidingBlocks", short.class);
+            SET_SPECIAL_COLLIDING_BLOCKS = lookup.findSetter(LevelChunkSection.class, "specialCollidingBlocks", short.class);
             TICKING_BLOCKS = ReflectionUtils.getField(LevelChunkSection.class, "tickingBlocks");
             
             Field fluidStateCount = ReflectionUtils.getFieldOrNull(LevelChunkSection.class, "fluidStateCount");
